Test Updates - Now supports simulating users
I added testing support that simulates a user adding inputs. Originally, the test was failing because the ForeignCourseStudent table creates an entry every time a Foreign Course is made, with the Student data originating from the current student's data. To fix this, I created a simulation in which the tests are operating as if they were a user creating inputs. This will be useful in the future.