Debugging
I fixed the views/tests so that the controller should work. The failures stemmed from 2 issues - Number one, the usage of Student.Where instead of Student.find_by_id, and Number two: Some of the tests on this branch had not been updated to mesh well with the implementation of roles. I've fixed this, so everything should be good now.